requirements.txt
PyYAML~=6.0.2
marshmallow_dataclass~=8.7.1
schedule~=1.2.2
requests~=2.32.3
# ZODB~=6.0
# simple-term-menu~=1.6.6
sqlite-web~=0.6.4

# Optional for development
Flask~=3.1.0
rich~=13.9.4
